% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_answer_option_to_question.R
\name{add_answer_option_to_question}
\alias{add_answer_option_to_question}
\title{Add an answer option to a question}
\usage{
add_answer_option_to_question(question, ...)
}
\arguments{
\item{question}{The limonaid Question object}

\item{...}{Options that are passed on to the Question's \code{add_answer_option()}
method.}
}
\value{
The question object.
}
\description{
This is a convenience function that allows you to add an answer option to
a question object.
}
\examples{
myQuestion <-
  limonaid::Question$new(
    code = 'myQuestion',
    type='radio'
  ) |>
  add_answer_option_to_question(
    code = 1,
    optionTexts = "First option"
  ) |>
  add_answer_option_to_question(
    code = 2,
    optionTexts = "Second option"
  );
}
