% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_lsdf_rows.R
\name{append_lsdf_rows}
\alias{append_lsdf_rows}
\title{A home-rolled version of plyr::rbind.fill}
\usage{
append_lsdf_rows(data, row)
}
\arguments{
\item{data}{The first dataframe.}

\item{row}{The second dataframe.}
}
\value{
A merged dataframe.
}
\description{
This is used when creating dataframes for TSV exports.
}
\examples{
limonaid::append_lsdf_rows(mtcars, iris);
}
