% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limer_base64_to_df.R
\name{limer_base64_to_df}
\alias{limer_base64_to_df}
\title{Convert base64 encoded data to a data frame}
\usage{
limer_base64_to_df(
  x,
  encoding = NULL,
  iconvArgs = list(from = "UTF-8", to = "UTF-8")
)
}
\arguments{
\item{x}{\dots}

\item{encoding}{Either \code{NULL} or an encoding to pass to \code{\link[=textConnection]{textConnection()}}.}

\item{iconvArgs}{Arguments to pass to [base::iconv().}
}
\description{
This function converts raw base64 results into a data frame.
}
\examples{
\dontrun{
limer_base64_to_df()
}

}
