% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limer_get_participant_property.R
\name{limer_get_participant_property}
\alias{limer_get_participant_property}
\title{Get a participant property from a LimeSurvey survey}
\usage{
limer_get_participant_property(
  iSurveyID,
  aTokenQueryProperties,
  aTokenProperties
)
}
\arguments{
\item{iSurveyID}{\dots}

\item{aTokenQueryProperties}{\dots}

\item{aTokenProperties}{\dots}
}
\description{
This function exports and downloads a participant property from a LimeSurvey survey.
}
\examples{
\dontrun{
limer_get_participant_property(
  iSurveyID = 12345,
  aTokenQueryProperties = 1,
  aTokenProperties = list("attribute_1")
);
}
}
