% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_parse_data_import_script.R
\name{ls_parse_data_import_script}
\alias{ls_parse_data_import_script}
\title{Extract specific code bits from LimeSurvey data import R script}
\usage{
ls_parse_data_import_script(
  scriptfile = NULL,
  scriptEncoding = limonaid::opts$get("encoding"),
  silent = limonaid::opts$get("silent")
)
}
\arguments{
\item{scriptfile}{The path and filename of the script file.}

\item{scriptEncoding}{The encoding of the script file; can be used
to override the setting in the \code{limonaid} options (i.e. in \code{opts}) in the
\code{encoding} field (the default value is "\code{UTF-8}").}

\item{silent}{Whether to be silent or verbose ('chatty').}
}
\value{
A list with four components.
}
\description{
This function extracts specific code bits from the LimeSurvey data
import R script, which can then be applied to imported data using
\code{\link[=ls_apply_script_bits]{ls_apply_script_bits()}}, for example to update variable names, set
labels, etc.
}
