% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_tsv_get_group_rows.R
\name{ls_tsv_get_group_rows}
\alias{ls_tsv_get_group_rows}
\title{Get all group rows from a LimeSurvey survey dataframe}
\usage{
ls_tsv_get_group_rows(data)
}
\arguments{
\item{data}{The LimeSurvey survey dataframe.}
}
\value{
A dataframe with the rows.
}
\description{
Get all group rows from a LimeSurvey survey dataframe
}
\examples{
### Add
}
