% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_tsv_get_rows.R
\name{ls_tsv_get_rows}
\alias{ls_tsv_get_rows}
\title{Display rows from a LimeSurvey dataframe that meet a criterion}
\usage{
ls_tsv_get_rows(data, ...)
}
\arguments{
\item{data}{The datafram.}

\item{...}{For now, one column/value pair (the criterion).}
}
\value{
The rows, passed through \code{\link[=ls_tsv_rows]{ls_tsv_rows()}}.
}
\description{
Display rows from a LimeSurvey dataframe that meet a criterion
}
\examples{
### Add later
}
