% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_write_tsv.R
\name{ls_write_tsv}
\alias{ls_write_tsv}
\title{Write a data frame to a LimeSurvey Tab Separated Values file}
\usage{
ls_write_tsv(
  data,
  file,
  encoding = limonaid::opts$get("encoding"),
  preventOverwriting = limonaid::opts$get("preventOverwriting"),
  silent = limonaid::opts$get("silent")
)
}
\arguments{
\item{data}{The dataframe to write.}

\item{file}{The file to write to.}

\item{encoding}{The encoding to write to.}

\item{preventOverwriting}{Whether to prevent overwriting, should the
target file exist, already.}

\item{silent}{Whether to be silent or chatty.}
}
\value{
The dataframe, adapted for writing, invisibly.
}
\description{
Write a data frame to a LimeSurvey Tab Separated Values file
}
\examples{
### Add example once something is available.
}
