% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsdf_for_language.R
\name{lsdf_for_language}
\alias{lsdf_for_language}
\title{Produce the dataframe containing the survey for one language}
\usage{
lsdf_for_language(
  language,
  groups,
  exportGroupIdMapping,
  exportQuestionIdMapping,
  backupLanguage,
  silent = limonaid::opts$get("silent")
)
}
\arguments{
\item{language}{The language for which to produce the data frame.}

\item{groups}{The groups object in the Survey object.}

\item{exportGroupIdMapping, exportQuestionIdMapping}{Used to map Survey
object identifier onto the identifier model used in the LimeSurvey TSV.}

\item{backupLanguage}{The language to get content from if not available in
the primary language}

\item{silent}{Whether to be silent or chatty.}
}
\value{
Invisibly, the \code{Survey} object.
}
\description{
This is used when exporting surveys to LimeSurvey's TSV format.
}
