% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limer_mail_registered_participant.R
\name{mail_registered_participant}
\alias{mail_registered_participant}
\title{Mail registered participant}
\usage{
mail_registered_participant(iSurveyID, tid)
}
\arguments{
\item{iSurveyID}{\dots}

\item{tid}{\dots}
}
\description{
This function was
adapted by Gjalt-Jorn Peters from a function originally written by Andrew
Heiss.
}
\examples{
\dontrun{
limonaid::mail_registered_participant(iSurveyID = 123456, tid = 2)
}
}
