% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processLimeSurveyDropouts.R
\name{processLimeSurveyDropouts}
\alias{processLimeSurveyDropouts}
\title{Process LimeSurvey dropouts}
\usage{
processLimeSurveyDropouts(lastpage, pagenames = NULL, relevantPagenames = NULL)
}
\arguments{
\item{lastpage}{A vector with the 'lastpage' variable as LimeSurvey stores
it (an integer denoting the last page a participant visited, in other words,
where they dropped out).}

\item{pagenames}{Optional: names for each page.}

\item{relevantPagenames}{Optional: the names of those pages that should be
included.}
}
\value{
A list with information about the dropout, including plots.
}
\description{
This function makes it easy to parse the dropouts from a LimeSurvey
questionnaire.
}
\details{
This will be described more in detail in a forthcoming publications.
}
\examples{
limonaid::processLimeSurveyDropouts(c(1,2,1,1,2,3,2,2,3,2,1));

}
