% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_survey.R
\name{Survey}
\alias{Survey}
\title{R6 Class representing a LimeSurvey survey}
\description{
R6 Class representing a LimeSurvey survey

R6 Class representing a LimeSurvey survey
}
\details{
A survey has at least a title and a primary language.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{The title of the survey in the primary language.}

\item{\code{primaryLanguage}}{The primary language of the survey.}

\item{\code{groups}}{The groups in the survey.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_group_ids}}{A list of all group ids.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Survey$new()}}
\item \href{#method-add_group}{\code{Survey$add_group()}}
\item \href{#method-add_question}{\code{Survey$add_question()}}
\item \href{#method-clone}{\code{Survey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new survey object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$new(title, primaryLanguage = "en")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title.}

\item{\code{primaryLanguage}}{The primary language.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Survey} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_group"></a>}}
\if{latex}{\out{\hypertarget{method-add_group}{}}}
\subsection{Method \code{add_group()}}{
Add a group to a survey object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_group(titles, descriptions = "", relevance = 1, random_group = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{titles}}{The group's title, either as a named character vector
where each element is the group title in a different language, and
every element's name is the language code; or as a single character
value, in which case the survey's primary language is used.}

\item{\code{descriptions}}{The group description, either as a named character
vector where each element is the group description in a different
language, and every element's name is the language code; or as a single
character value, in which case the survey's primary language is used.}

\item{\code{relevance}}{The group's relevance equation.}

\item{\code{random_group}}{The group's randomization group.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly, the \code{Survey} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_question"></a>}}
\if{latex}{\out{\hypertarget{method-add_question}{}}}
\subsection{Method \code{add_question()}}{
Add a question to a survey object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_question(groupId, code, type = NULL, lsType = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groupId}}{The id of the group to add the question to.}

\item{\code{code}}{The question code.}

\item{\code{type}}{The question type.}

\item{\code{lsType}}{The question type, as LimeSurvey question type.}

\item{\code{...}}{Additional arguments are used to create the Question using
\code{Question$new}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly, the \code{Survey} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
