% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{Transpose a data frame}
\usage{
transpose_df(x)
}
\arguments{
\item{x}{The data frame.}
}
\value{
A list.
}
\description{
Returns a list of lists, where each list contains a row.
}
\examples{
limonaid::transpose_df(
  mtcars[1:3, 1:3]
);
}
