% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limonaid-package.R
\docType{package}
\name{limonaid-package}
\alias{limonaid-package}
\alias{limonaid}
\title{limonaid-package}
\description{
Working With LimeSurvey Surveys and Responses
}
\details{
LimeSurvey is Free/Libre Open Source Software for
the development and administrations of online studies, using
sophisticated tailoring capabilities to support multiple study
designs. This package supports programmatic creation of surveys
that can then be imported into LimeSurvey, as well as
userfriendly import of responses from LimeSurvey studies.
}
\author{
Gjalt-Jorn Ygram Peters \href{mailto:gjalt-jorn@behaviorchange.eu}{gjalt-jorn@behaviorchange.eu}
}
\keyword{package}
