% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_question.R
\name{Question}
\alias{Question}
\title{R6 Class representing a LimeSurvey question}
\description{
R6 Class representing a LimeSurvey question

R6 Class representing a LimeSurvey question
}
\details{
A question has at least a code and a primary language.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{The code of the question.}

\item{\code{id}}{The identifier of the question (a unique number in a survey)}

\item{\code{type}}{The question type.}

\item{\code{lsType}}{The question type in LimeSurvey's format.}

\item{\code{questionTexts}}{The question text(s) in all languages.}

\item{\code{helpTexts}}{The question help text(s) in all languages.}

\item{\code{relevance}}{The relevance.}

\item{\code{primaryLanguage}}{The primary language of the question.}

\item{\code{answerOptions}}{The answer options in the question.}

\item{\code{subQuestions}}{The subquestions in the question.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Question$new()}}
\item \href{#method-add_answer_option}{\code{Question$add_answer_option()}}
\item \href{#method-clone}{\code{Question$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new question object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$new(
  code,
  type = NULL,
  lsType = NULL,
  id = NULL,
  questionTexts = "",
  helpTexts = "",
  relevance = 1,
  primaryLanguage = "en"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{The question code.}

\item{\code{type}}{The human-readable question type.}

\item{\code{lsType}}{The type as LimeSurvey type (see
https://manual.limesurvey.org/Question_object_types).}

\item{\code{id}}{The identifier of the question (in a survey).}

\item{\code{questionTexts}}{The question text(s).}

\item{\code{helpTexts}}{The help text(s).}

\item{\code{relevance}}{The group's relevance equation.}

\item{\code{primaryLanguage}}{The question's primary language.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Survey} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_answer_option"></a>}}
\if{latex}{\out{\hypertarget{method-add_answer_option}{}}}
\subsection{Method \code{add_answer_option()}}{
Add an answer option to a question.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$add_answer_option(code, optionTexts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{The answer option code.}

\item{\code{optionTexts}}{The answer option text(s).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible, the question object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
