% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_survey.R
\name{Survey}
\alias{Survey}
\title{R6 Class representing a LimeSurvey survey}
\description{
R6 Class representing a LimeSurvey survey

R6 Class representing a LimeSurvey survey
}
\details{
A survey has at least a title and a primary language.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{The title of the survey in the primary language.}

\item{\code{primaryLanguage}}{The primary language of the survey.}

\item{\code{groups}}{The groups in the survey.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Survey$new()}}
\item \href{#method-add_group}{\code{Survey$add_group()}}
\item \href{#method-add_question}{\code{Survey$add_question()}}
\item \href{#method-clone}{\code{Survey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new survey object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$new(title, primaryLanguage)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{Title.}

\item{\code{primaryLanguage}}{The primary language.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Survey} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_group"></a>}}
\if{latex}{\out{\hypertarget{method-add_group}{}}}
\subsection{Method \code{add_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_group(
  title,
  description = "",
  relevance = 1,
  random_group = "",
  language = self$primaryLanguage
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_question"></a>}}
\if{latex}{\out{\hypertarget{method-add_question}{}}}
\subsection{Method \code{add_question()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_question(
  group,
  code,
  text = "",
  relevance = 1,
  language = self$primaryLanguage
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
